Bubble Sort
step-1: You compare the first item with the second. If the first item is bigger than the second item. swapping occurs.
step-2:And then compare further, second with third item. if second item is bigger than the third, we swap them. Else remain as it is.
step-3:Compare until we hit the last element of the array.
step-4: Look at the inner loop in the code below.
step-5: Now we repeat again process, starting from the last item of the array. 
step-6: and then we move backward inside the outer loop.
Loop Continues until array is sorted.
Main goal is to sort the array from shortest number to highest number.